third_party/WebKit/LayoutTests/fast/block/margin-collapse/webkit-margin-collapse-separate-position-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=109956 -webkit-margin-collapse: separate doesn't work correctly for before margins. The test basically creates an empty block inside a container with a margin that collapses with children. The empty block has the margins set to separate so its height should be 0px. The container is not selfcollapsing so it should have a height of 10px+10px+2px=22px.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("zero_height").offsetHeight is 0
PASS document.getElementById("22px_height").offsetHeight is 22
PASS successfullyParsed is true

TEST COMPLETE
